refactor: update trie update tests and remove obsolete test
- Replaced the obsolete `trie_fork_test` with `trie_updates_test` to focus on validating trie updates in canonical blocks.
- Adjusted the `ProduceBlockWithTransactionsViaEngineAPI` to ensure proper handling of trie updates.
- Updated the `parent_beacon_block_root` in transaction payload attributes to align with the latest block structure.
- Enhanced test configurations to ensure compatibility with the Cancun fork and improve testing accuracy.
